refactor card concepts
A `Card` instance is always a handle to some other card that you wish to
interact with.

A `ChildCard` is a handle to a card that is embedded inside your own
page (and thus it has an iframe).

A `ParentCard` is a handle to the page that contains you.

A `PeerCard` (not fully implemented yet) is neither -- you are both
descendants of the same root card.