fix width measurement
We need to use parent instead of offsetParent here.

In the most general case, even `parent` is too simplistic, but we're not
dealing with the general case. The containing environment creates an
arbitrary element to put our card inside, but our card element
itself (the iframe) is not something users are supposed to be trying to
position, so we can assume it is statically positioned and that it's
parent element provides the correct block context.